rename specHelpers to withContext
