+++ test/ReactViews/Tools/ItemSearchTool/ItemSearchToolSpec.tsx
@@ -22,7 +22,7 @@ import ItemSearchTool, {
 import Loading from "../../../../lib/ReactViews/Tools/ItemSearchTool/Loading";
 import SearchForm from "../../../../lib/ReactViews/Tools/ItemSearchTool/SearchForm";
 import SearchResults from "../../../../lib/ReactViews/Tools/ItemSearchTool/SearchResults";
-import { withThemeContext } from "../../specHelpers";
+import { withThemeContext } from "../../withContext";
 import MockSearchableItem from "./MockSearchableItem";
 import i18next from "i18next";
