+++ test/ReactViews/WelcomeMessageSpec.tsx
@@ -4,7 +4,7 @@ import { act } from "react-dom/test-utils";
 import Terria from "../../lib/Models/Terria";
 import ViewState from "../../lib/ReactViewModels/ViewState";
 import { WelcomeMessagePure } from "../../lib/ReactViews/WelcomeMessage/WelcomeMessage";
-import { createWithContexts } from "./SpecHelpers";
+import { createWithContexts } from "./withContext";
 const WelcomeMessage: any = require("../../lib/ReactViews/WelcomeMessage/WelcomeMessage")
   .default;
